requirements.txt
discord.py>=2.3.0
motor>=3.3.0
Pillow>=10.0.0
aiohttp>=3.8.5 